refactor(earn): move useApyWindow hook to src/hooks/earn - JUM-839
