+++ src/components/EarnDetails/EarnDetailsIntro.tsx
@@ -1,22 +1,22 @@
 'use client';
 
+import { useApyWindow } from '@/hooks/earn/useApyWindow';
+import { useZapEarnOpportunitySlugStorage } from '@/providers/hooks';
+import useMediaQuery from '@mui/material/useMediaQuery';
+import { formatDistance } from 'date-fns';
 import type { FC } from 'react';
 import { useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
 import type { EarnOpportunityWithLatestAnalytics } from 'src/types/jumper-backend';
-import { EarnCard } from '../Cards/EarnCard/EarnCard';
 import { Badge } from '../Badge/Badge';
 import { BadgeSize, BadgeVariant } from '../Badge/Badge.styles';
-import useMediaQuery from '@mui/material/useMediaQuery';
+import { EarnCard } from '../Cards/EarnCard/EarnCard';
 import { ProtocolCard } from '../Cards/ProtocolCard/ProtocolCard';
 import {
   EarnDetailsColumnFlexContainer,
   EarnDetailsRowFlexContainer,
 } from './EarnDetails.styles';
 import { EarnDetailsActions } from './EarnDetailsActions';
-import { formatDistance } from 'date-fns';
-import { useTranslation } from 'react-i18next';
-import { useZapEarnOpportunitySlugStorage } from '@/providers/hooks';
-import { useApyWindow } from '@/components/EarnFilterBar/components/useApyWindow';
 
 interface EarnDetailsIntroProps {
   data: EarnOpportunityWithLatestAnalytics;
